mazeGame runs, graphics all decode, but text rendering still absent
