+++ src/lib/mod.rs
@@ -6,4 +6,3 @@ extern crate candid;
 //extern crate candid_derive;
 
 pub mod types;
-pub mod render;
